v1/requirements.txt
fastapi>=0.95
uvicorn>=0.22
pydantic>=2.6
requests>=2.31
sqlalchemy>=2.0,<2.1
asyncpg>=0.29
pymilvus>=2.1
minio>=7.1
kafka-python>=2.0
pdfplumber>=0.7
pytesseract>=0.3
sentence-transformers>=2.2
opentelemetry-api>=1.23
opentelemetry-sdk>=1.23
prometheus-client>=0.16
